Move UrlUnicodeNormalizer to impl module
