--- autofill/autofill-api/src/main/java/com/duckduckgo/autofill/api/encoding/UrlUnicodeNormalizer.kt
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.autofill.api.encoding
-
-interface UrlUnicodeNormalizer {
-    fun normalizeAscii(url: String?): String?
-    fun normalizeUnicode(url: String?): String?
-}
