+++ autofill/autofill-impl/src/main/java/com/duckduckgo/autofill/impl/urlmatcher/AutofillDomainNameUrlMatcher.kt
@@ -19,7 +19,7 @@ package com.duckduckgo.autofill.impl.urlmatcher
 import androidx.core.net.toUri
 import com.duckduckgo.app.global.extractDomain
 import com.duckduckgo.app.global.normalizeScheme
-import com.duckduckgo.autofill.api.encoding.UrlUnicodeNormalizer
+import com.duckduckgo.autofill.impl.encoding.UrlUnicodeNormalizer
 import com.duckduckgo.autofill.impl.urlmatcher.AutofillUrlMatcher.ExtractedUrlParts
 import javax.inject.Inject
 import okhttp3.HttpUrl.Companion.toHttpUrl
