+++ autofill/autofill-impl/src/test/java/com/duckduckgo/autofill/impl/sharedcreds/AppleShareableCredentialsTest.kt
@@ -1,9 +1,9 @@
 package com.duckduckgo.autofill.impl.sharedcreds
 
 import com.duckduckgo.app.CoroutineTestRule
-import com.duckduckgo.autofill.api.TestUrlUnicodeNormalizer
 import com.duckduckgo.autofill.api.domain.app.LoginCredentials
 import com.duckduckgo.autofill.api.store.AutofillStore
+import com.duckduckgo.autofill.impl.encoding.TestUrlUnicodeNormalizer
 import com.duckduckgo.autofill.impl.sharedcreds.SharedCredentialsParser.SharedCredentialConfig
 import com.duckduckgo.autofill.impl.urlmatcher.AutofillDomainNameUrlMatcher
 import com.duckduckgo.autofill.impl.urlmatcher.AutofillUrlMatcher.ExtractedUrlParts
