+++ autofill/autofill-impl/src/test/java/com/duckduckgo/autofill/impl/ui/credential/selecting/RealAutofillSelectCredentialsGrouperTest.kt
@@ -17,8 +17,8 @@
 package com.duckduckgo.autofill.impl.ui.credential.selecting
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.duckduckgo.autofill.api.TestUrlUnicodeNormalizer
 import com.duckduckgo.autofill.api.domain.app.LoginCredentials
+import com.duckduckgo.autofill.impl.encoding.TestUrlUnicodeNormalizer
 import com.duckduckgo.autofill.impl.ui.credential.management.sorting.CredentialListSorterByTitleAndDomain
 import com.duckduckgo.autofill.impl.ui.credential.selecting.AutofillSelectCredentialsGrouper.Groups
 import com.duckduckgo.autofill.impl.urlmatcher.AutofillDomainNameUrlMatcher
